buygift-auto-buyer/requirements.txt
selenium>=4.10.0
PyYAML>=6.0.1
webdriver-manager>=4.0.0
